pyproject.toml
[tool.poetry]
name = "live-flight"
version = "0.1.0"
description = "Shows the closest live flight to your current location using the OpenSky Network API."
authors = ["Samuel Favarin <samuel.favarin@databuildcompany.com>"]
readme = "README.md"
packages = [{ include = "live_flight" }]

[tool.poetry.dependencies]
python = "^3.10"
opensky-api = { git = "https://github.com/openskynetwork/opensky-api.git", subdirectory = "python" }
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"

[tool.poetry.scripts]
live-flight = "live_flight.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=live_flight --cov-report=term-missing"

[tool.coverage.run]
branch = true
source = ["live_flight"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
